Fix koditool test failure and resolve workspace warnings
- Created docs/specs/0003-fix-koditool-tests.md to document the fix strategy.
- Modified RpcClient in koditool to support deterministic RNG seeding via an optional seed field and `with_seed` method.
- Updated koditool/tests/kodi_helper_test.rs to use a fixed seed, ensuring deterministic behavior.
- Resolved compilation and lint warnings across the workspace:
  - Removed unused imports in koditool and tv_mode_web.
  - Suppressed dead code warnings in koditool by prefixing with underscore or using #[allow(dead_code)].
  - Updated rand::thread_rng() to rand::rng() in tv_mode_web.
  - Fixed private interface leak in tv_mode_web/src/routes/jukectl.rs.
  - Refactored dupehunter into separate lib and bin to resolve multiple build targets warning.
  - Removed unnecessary parentheses in jukeingest.
- Verified all tests pass via `make test`.
