+++ tv_mode_web/src/app_state.rs
@@ -12,7 +12,7 @@ use rocket::serde::Deserialize;
 use rocket::serde::Serialize;
 use std::collections::BTreeMap;
 use std::collections::HashMap;
-use std::time::{Duration, SystemTime, UNIX_EPOCH};
+use std::time::{SystemTime, UNIX_EPOCH};
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct ShowMappings {
